feat: implementar sistema de gerenciamento de imagens com ContentType
Adiciona novo value object ContentType para validação de tipos de conteúdo e implementa sistema completo de upload, armazenamento e exclusão de imagens:
- Criado ContentType value object com validação e testes
- Implementado processamento de imagens no ProductController
- Adicionado handler para criação de produtos com múltiplas imagens
- Criado sistema de exclusão de imagens (dispatcher, job, scheduler)
- Atualizada migration para nova estrutura da tabela Image
- Refatorados testes e serviços de infraestrutura
- Removida migration antiga de relacionamento Product-Image

+++ src/Riber.Application/Exceptions/BadRequestException.cs
@@ -3,4 +3,4 @@ using System.Net;
 namespace Riber.Application.Exceptions;
 
 public sealed class BadRequestException(string message) 
-    : ApplicationException(message, (int)HttpStatusCode.BadRequest);
+    : ApplicationException(message, HttpStatusCode.BadRequest);
